docs: revert license to GPLv2; add file headers; add authorship details
This commit corrects an accidental and unauthorized change of the project's license from GNU GPL v2 to GNU GPL v3.

The original FERS project is licensed under the terms of the GNU General Public License v2.0 only. As this fork is a derivative work, it must adhere to the original license.